requirements.txt
fastapi
uvicorn[standard]
boto3
PyPDF2
python-dotenv
sentence-transformers
faiss-cpu